+++ frontend/src/ast/declaration/DeclarationList.c
@@ -15,7 +15,7 @@ bool parse_declaration_list(ParserState* s, DeclarationList* res) {
         return false;
     }
 
-    size_t alloc_size = res->len;
+    uint32_t alloc_size = res->len;
     while (is_declaration(s)) {
         if (res->len == alloc_size) {
             mycc_grow_alloc((void**)&res->decls, &alloc_size, sizeof *res->decls);
@@ -33,7 +33,7 @@ bool parse_declaration_list(ParserState* s, DeclarationList* res) {
 }
 
 void DeclarationList_free(DeclarationList* l) {
-    for (size_t i = 0; i < l->len; ++i) {
+    for (uint32_t i = 0; i < l->len; ++i) {
         Declaration_free_children(&l->decls[i]);
     }
     mycc_free(l->decls);
